move stuff back to Source::SourceLocator
