+++ lib/reek/source/source_code.rb
@@ -4,7 +4,6 @@ Reek::CLI::Silencer.silently do
 end
 require_relative '../tree_dresser'
 require_relative '../ast/node'
-require_relative '../../../lib/reek/source/source_path'
 
 module Reek
   module Source
@@ -42,7 +41,7 @@ module Reek
         case source
         when File       then new(source.read, source.path)
         when IO         then new(source.readlines.join, IO_IDENTIFIER)
-        when SourcePath then new(source.read, source.to_s)
+        when SourceLocator then new(source.read, source.to_s)
         when String     then new(source, STRING_IDENTIFIER)
         end
       end
