+++ spec/reek/spec/should_reek_spec.rb
@@ -22,9 +22,9 @@ RSpec.describe Reek::Spec::ShouldReek do
   end
 
   describe 'checking code in a File' do
-    let(:clean_file) { Reek::Source::SourcePath.new(SAMPLES_PATH.join('three_clean_files/clean_one.rb')) }
-    let(:smelly_file) { Reek::Source::SourcePath.new(SAMPLES_PATH.join('two_smelly_files/dirty_one.rb')) }
-    let(:masked_file) { Reek::Source::SourcePath.new(SAMPLES_PATH.join('clean_due_to_masking/dirty_one.rb')) }
+    let(:clean_file) { Reek::Source::SourceLocator.new(SAMPLES_PATH.join('three_clean_files/clean_one.rb')) }
+    let(:smelly_file) { Reek::Source::SourceLocator.new(SAMPLES_PATH.join('two_smelly_files/dirty_one.rb')) }
+    let(:masked_file) { Reek::Source::SourceLocator.new(SAMPLES_PATH.join('clean_due_to_masking/dirty_one.rb')) }
 
     context 'matcher without masking' do
       let(:matcher) { Reek::Spec::ShouldReek.new }
@@ -47,7 +47,7 @@ RSpec.describe Reek::Spec::ShouldReek do
       let(:path) { SAMPLES_PATH.join('clean_due_to_masking/masked.reek') }
       let(:configuration) { test_configuration_for(path) }
       let(:matcher) { Reek::Spec::ShouldReek.new(configuration: configuration) }
-      let(:masked_file) { Reek::Source::SourcePath.new(SAMPLES_PATH.join('clean_due_to_masking/dirty_one.rb')) }
+      let(:masked_file) { Reek::Source::SourceLocator.new(SAMPLES_PATH.join('clean_due_to_masking/dirty_one.rb')) }
 
       it 'masks smells using the relevant configuration' do
         expect(matcher.matches?(masked_file)).to be_falsey
